fix usdt precision

+++ routes/json.js
@@ -1363,8 +1363,8 @@ router.get('/oracle/:tracker/:from-:to', (req, res, next) => {
     let bnbRates = getNestedObject(oracle_report, [Object.keys(oracle_report)[0],'bnb']) != undefined ? getNestedObject(oracle_report, [Object.keys(oracle_report)[0],'bnb']) : 0;
     let bscRates = getNestedObject(oracle_report, [Object.keys(oracle_report)[0],'bsc']) != undefined ? getNestedObject(oracle_report, [Object.keys(oracle_report)[0],'bsc']) : 0;
     console.log("ETH currency format: "+CurrencyJS(ethRates, { fromCents: false, precision: 16, separator: ',' }).format()); // "123456" => "123456.00" =>  "123,456.00"
-    var rateUSDformatCurrency = CurrencyJS(usdRates, { symbol: '', fromCents: false, precision: 2, separator: ',' }).format(); // "123456" => "123456.00" =>  "123,456.00"
-    var rateUSDTformatCurrency = CurrencyJS(usdtRates, { symbol: '', fromCents: false, precision: 8, separator: ',' }).format(); // "123456" => "123456.00" =>  "123,456.00"
+    var rateUSDformatCurrency = CurrencyJS(usdRates, { symbol: '', fromCents: false, precision: 16, separator: ',' }).format(); // "123456" => "123456.00" =>  "123,456.00"
+    var rateUSDTformatCurrency = CurrencyJS(usdtRates, { symbol: '', fromCents: false, precision: 16, separator: ',' }).format(); // "123456" => "123456.00" =>  "123,456.00"
     var rateBTCformatCurrency = CurrencyJS(btcRates, { symbol: '', fromCents: false, precision: 8, separator: ',' }).format(); // "123456" => "123,456.00" ? Ξ Ł
     //ŁΞ₿₮'$',
     var rateLTCformatCurrency = CurrencyJS(ltcRates, { symbol: '', fromCents: false, precision: 8, separator: ',' }).format(); // "123456" => "123,456.00" ? Ξ Ł
